backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Google AI (for embeddings and Gemini fallback)
google-generativeai==0.4.0

# OpenRouter (uses OpenAI SDK compatibility)
openai>=1.0.0

# Vector Store
# Note: chromadb 0.4.x requires numpy<2.0 for compatibility
chromadb>=0.5.0
numpy<2.0

# Document Processing - PDF & Image OCR
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.16.3

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Production Auth & Async
PyJWT==2.8.0
redis==5.0.1
rq==1.15.1

# Development & Testing
pytest==7.4.4
httpx==0.26.0

# Notes:
# - Tesseract OCR must be installed separately on the system
# - Windows: https://github.com/UB-Mannheim/tesseract/wiki
# - Mac: brew install tesseract  
# - Linux: sudo apt-get install tesseract-ocr
# - pdf2image requires poppler-utils (Linux/Mac) or poppler for Windows
